 requirements.txt
dotenv
langchain-openai
pandas
pinecone-client
pinecone-text
tabulate
tiktoken
google-cloud-documentai
